+++ playwright/tests/sponsor-logo.spec.ts
@@ -1,7 +1,7 @@
 import { test } from '@playwright/test';
-import { cmpAcceptAll } from '../../lib/cmp';
-import { loadPage } from '../../lib/load-page';
-import { getStage, getTestUrl, waitForSlot } from '../../lib/util';
+import { cmpAcceptAll } from '../lib/cmp';
+import { loadPage } from '../lib/load-page';
+import { getStage, getTestUrl, waitForSlot } from '../lib/util';
 
 test.describe('sponsorshipLogo', () => {
 	test('sponsor logo ad is correctly filled in thrasher fixture', async ({