+++ src/lib/remove-slots.ts
@@ -1,7 +1,5 @@
 import { log } from '@guardian/libs';
 import { once } from 'lodash-es';
-import { isInVariantSynchronous } from 'experiments/ab';
-import { elementsManager } from 'experiments/tests/elements-manager';
 import { dfpEnv } from '../dfp/dfp-env';
 import fastdom from './fastdom-promise';
 
@@ -25,17 +23,12 @@ const filterDisabledNodes = (nodes: Element[]) => nodes.filter(isDisabled);
 const removeNodes = (nodes: Element[]): Promise<void> =>
 	fastdom.mutate(() =>
 		nodes.forEach((node) => {
-			log('commercial', `Removing ad slot ${node.id}`);
+			log('commercial', `Removing ad slot: ${node.id}`);
 			node.remove();
 		}),
 	);
 
 const removeSlots = (): Promise<void> => {
-	// Don't collapse slots when in the Elements Manager AB test variant.
-	if (isInVariantSynchronous(elementsManager, 'variant')) {
-		return Promise.resolve();
-	}
-
 	return removeNodes(selectNodes());
 };